DB 수정 완료

+++ src/main/java/com/Capstone/EduX/Entity/ExamQuestion.java
@@ -12,5 +12,5 @@ public class ExamQuestion {
     private String question;
     private String answer;
     private String type;
-    private float questionScore;
+    private Integer questionScore;
 }
